fix(onboard): route portable gateway reachability probe
